pyproject.toml
[tool.poetry]
name = "repl_python3_WhatsApp-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
brotli = "^1.0.9"
htmlparser = "^0.0.2"
importlib-metadata = "^4.5.0"
monotonic = "^1.6"
ordereddict = "^1.1"
python-ntlm = "^1.1.0"
zipp = "^3.4.1"
cryptography = "^3.4.7"
keyring = "^23.0.1"
lockfile = "^0.12.2"
lxml = "^4.6.3"
tornado = "^6.1"
flask = "^2.0.1"
twilio = "^6.59.1"
aiohttp = "^3.7.4"
markdownify = "^0.9.0"
bs4 = "^0.0.1"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
